Make brms model print list

make_brms_mdl_print_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make brms model print list. The function returns Brms model print (a list).

make_brms_mdl_print_ls(
  mdl_ls,
  label_stub_1L_chr,
  caption_1L_chr,
  output_type_1L_chr = "PDF",
  digits_1L_dbl = 2,
  big_mark_1L_chr = " "
)

Arguments

mdl_ls
Model (a list)

label_stub_1L_chr
Label stub (a character vector of length one)

caption_1L_chr
Caption (a character vector of length one)

output_type_1L_chr
Output type (a character vector of length one), Default: 'PDF'

digits_1L_dbl
Digits (a double vector of length one), Default: 2

big_mark_1L_chr
Big mark (a character vector of length one), Default: ' '

Value

Brms model print (a list)