Make preferred models vector

make_prefd_mdls_vec() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make preferred models vector. The function returns Preferred models (a character vector).

make_prefd_mdls_vec(
  smry_of_sngl_predr_mdls_tb,
  choose_from_pfx_chr = c("GLM", "OLS"),
  mdl_types_lup = NULL
)

Arguments

smry_of_sngl_predr_mdls_tb
Summary of single predictor models (a tibble)

choose_from_pfx_chr
Choose from prefix (a character vector), Default: c("GLM", "OLS")

mdl_types_lup
Model types (a lookup table), Default: NULL

Value

Preferred models (a character vector)