Make results list

make_results_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make results list. The function returns Results (a list).

make_results_ls(
  spine_of_results_ls = NULL,
  abstract_args_ls = NULL,
  dv_ds_nm_and_url_chr = NULL,
  output_format_ls = NULL,
  params_ls_ls = NULL,
  path_params_ls = NULL,
  study_descs_ls = NULL,
  fn_ls = NULL,
  include_idx_int = NULL,
  var_nm_change_lup = NULL,
  ctgl_vars_regrouping_ls = NULL,
  sig_covars_some_predrs_mdls_tb = NULL,
  sig_thresh_covars_1L_chr = NULL,
  version_1L_chr = NULL
)

Arguments

spine_of_results_ls
Spine of results (a list), Default: NULL

abstract_args_ls
Abstract arguments (a list), Default: NULL

dv_ds_nm_and_url_chr
Dataverse dataset name and url (a character vector), Default: NULL

output_format_ls
Output format (a list), Default: NULL

params_ls_ls
Parameters (a list of lists), Default: NULL

path_params_ls
Path parameters (a list), Default: NULL

study_descs_ls
Study descriptions (a list), Default: NULL

fn_ls
Function list (a list of functions), Default: NULL

include_idx_int
Include index (an integer vector), Default: NULL

var_nm_change_lup
Variable name change (a lookup table), Default: NULL

ctgl_vars_regrouping_ls
Categorical variables regrouping (a list), Default: NULL

sig_covars_some_predrs_mdls_tb
Sig covariates some predictors models (a tibble), Default: NULL

sig_thresh_covars_1L_chr
Sig thresh covariates (a character vector of length one), Default: NULL

version_1L_chr
Version (a character vector of length one), Default: NULL

Value

Results (a list)