Make smry of model

make_smry_of_mdl() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make smry of model. The function returns Smry of one predr model (a tibble).

make_smry_of_mdl(
  data_tb,
  model_mdl,
  n_folds_1L_int = 10,
  dep_var_nm_1L_chr = "utl_total_w",
  start_1L_chr = NULL,
  tfmn_1L_chr = "NTF",
  predr_var_nm_1L_chr,
  covar_var_nms_chr = NA_character_,
  mdl_type_1L_chr = "OLS_NTF",
  mdl_types_lup = NULL,
  pred_type_1L_chr = NULL
)

Arguments

data_tb
Data (a tibble)

model_mdl
PARAM_DESCRIPTION

n_folds_1L_int
N folds (an integer vector of length one), Default: 10

dep_var_nm_1L_chr
Dep var name (a character vector of length one), Default: 'utl_total_w'

start_1L_chr
Start (a character vector of length one), Default: NULL

tfmn_1L_chr
Time series (a character vector of length one), Default: 'NTF'

predr_var_nm_1L_chr
Predr var name (a character vector of length one)

covar_var_nms_chr
Covar var names (a character vector), Default: 'NA'

mdl_type_1L_chr
Model type (a character vector of length one), Default: 'OLS_NTF'

mdl_types_lup
Model types (a lookup table), Default: NULL

pred_type_1L_chr
Pred type (a character vector of length one), Default: NULL

Value

Smry of one predr model (a tibble)