Make summary of model output

make_smry_of_mdl_outp() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make summary of model output. The function returns Summary of one predictor model (a tibble).

Usage

make_smry_of_mdl_outp(
  data_tb,
  model_mdl = NULL,
  folds_1L_int = 10,
  depnt_var_nm_1L_chr = "utl_total_w",
  start_1L_chr = NULL,
  tfmn_1L_chr = "NTF",
  predr_var_nm_1L_chr,
  covar_var_nms_chr = NA_character_,
  mdl_type_1L_chr = "OLS_NTF",
  mdl_types_lup = NULL,
  predn_type_1L_chr = NULL
)

Arguments

data_tb

Data (a tibble)

model_mdl

Model (a model), Default: NULL

folds_1L_int

Folds (an integer vector of length one), Default: 10

depnt_var_nm_1L_chr

Dependent variable name (a character vector of length one), Default: 'utl_total_w'

start_1L_chr

Start (a character vector of length one), Default: NULL

tfmn_1L_chr

Transformation (a character vector of length one), Default: 'NTF'

predr_var_nm_1L_chr

Predictor variable name (a character vector of length one)

covar_var_nms_chr

Covariate variable names (a character vector), Default: 'NA'

mdl_type_1L_chr

Model type (a character vector of length one), Default: 'OLS_NTF'

mdl_types_lup

Model types (a lookup table), Default: NULL

predn_type_1L_chr

Prediction type (a character vector of length one), Default: NULL

Value

Summary of one predictor model (a tibble)