Synthetic population replication dataset

A purely synthetic dataset, representative of the original study data, that can be used for replication runs of package algorithms.

replication_popl_tb

Format

An object of class data.frame with 1711 rows and 46 columns.

Details

A tibble

fkClientID

Unique Client Identifier

round

Round of Data Collection

c_p_diag_s

Primary Diagnosis

s_centre

Centre Name

c_clinical_staging_s

Clinical Stage

d_age

Age

d_agegroup

Age Group

d_gender

Gender

d_sex_birth_s

Sex at Birth

d_sexual_ori_s

Sexual Orientation

d_country_bir_s

Country Of Birth

d_ATSI

Aboriginal or Torres Strait Islander

d_english_home

Speaks English At Home

d_english_native

Native English Speaker

d_relation_s

Relationship Status

d_studying_working

Education and Employment Status

K6

Kessler Psychological Distress Scale (6 Dimension)

PHQ9

Patient Health Questionnaire

BADS

Behavioural Activation for Depression Scale

GAD7

Generalised Anxiety Disorder Scale

OASIS

Overall Anxiety Severity and Impairment Scale

SCARED

Screen for Child Anxiety Related Disorders

SOFAS

Social and Occupational Functioning Assessment Scale

Gender

Gender (Grouped)

Region

Region of Residence (Metropolitan or Regional)

CALD

Culturally And Linguistically Diverse

aqol6d_q1

Assessment of Quality of Life (6 Dimension) Question 1

aqol6d_q2

Assessment of Quality of Life (6 Dimension) Question 2

aqol6d_q3

Assessment of Quality of Life (6 Dimension) Question 3

aqol6d_q4

Assessment of Quality of Life (6 Dimension) Question 4

aqol6d_q5

Assessment of Quality of Life (6 Dimension) Question 5

aqol6d_q6

Assessment of Quality of Life (6 Dimension) Question 6

aqol6d_q7

Assessment of Quality of Life (6 Dimension) Question 7

aqol6d_q8

Assessment of Quality of Life (6 Dimension) Question 8

aqol6d_q9

Assessment of Quality of Life (6 Dimension) Question 9

aqol6d_q10

Assessment of Quality of Life (6 Dimension) Question 10

aqol6d_q11

Assessment of Quality of Life (6 Dimension) Question 11

aqol6d_q12

Assessment of Quality of Life (6 Dimension) Question 12

aqol6d_q13

Assessment of Quality of Life (6 Dimension) Question 13

aqol6d_q14

Assessment of Quality of Life (6 Dimension) Question 14

aqol6d_q15

Assessment of Quality of Life (6 Dimension) Question 15

aqol6d_q16

Assessment of Quality of Life (6 Dimension) Question 16

aqol6d_q17

Assessment of Quality of Life (6 Dimension) Question 17

aqol6d_q18

Assessment of Quality of Life (6 Dimension) Question 18

aqol6d_q19

Assessment of Quality of Life (6 Dimension) Question 19

aqol6d_q20

Assessment of Quality of Life (6 Dimension) Question 20