Write time series model plots

write_ts_mdl_plts() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write time series model plots. The function returns Model plots paths (a list).

write_ts_mdl_plts(
  brms_mdl,
  table_predn_mdl = NULL,
  tfd_data_tb,
  mdl_nm_1L_chr,
  path_to_write_to_1L_chr,
  depnt_var_nm_1L_chr = "utl_total_w",
  depnt_var_desc_1L_chr = "Utility score",
  predn_type_1L_chr = NULL,
  round_var_nm_1L_chr = "round",
  sd_dbl = NA_real_,
  sfx_1L_chr = " from table",
  tfmn_1L_chr = "NTF",
  units_1L_chr = "in",
  height_dbl = c(rep(6, 2), rep(5, 8)),
  width_dbl = c(rep(6, 2), rep(6, 8)),
  rsl_dbl = rep(300, 10),
  args_ls = NULL,
  seed_1L_dbl = 23456,
  utl_min_val_1L_dbl = -1
)

Arguments

brms_mdl
Bayesian regression models (a model)

table_predn_mdl
Table prediction (a model), Default: NULL

tfd_data_tb
Transformed data (a tibble)

mdl_nm_1L_chr
Model name (a character vector of length one)

path_to_write_to_1L_chr
Path to write to (a character vector of length one)

depnt_var_nm_1L_chr
Dependent variable name (a character vector of length one), Default: 'utl_total_w'

depnt_var_desc_1L_chr
Dependent variable description (a character vector of length one), Default: 'Utility score'

predn_type_1L_chr
Prediction type (a character vector of length one), Default: NULL

round_var_nm_1L_chr
Round variable name (a character vector of length one), Default: 'round'

sd_dbl
Standard deviation (a double vector), Default: NA

sfx_1L_chr
Suffix (a character vector of length one), Default: ' from table'

tfmn_1L_chr
Transformation (a character vector of length one), Default: 'NTF'

units_1L_chr
Units (a character vector of length one), Default: 'in'

height_dbl
Height (a double vector), Default: c(rep(6, 2), rep(5, 8))

width_dbl
Width (a double vector), Default: c(rep(6, 2), rep(6, 8))

rsl_dbl
Resolution (a double vector), Default: rep(300, 10)

args_ls
Arguments (a list), Default: NULL

seed_1L_dbl
Seed (a double vector of length one), Default: 23456

utl_min_val_1L_dbl
Utility minimum value (a double vector of length one), Default: -1

Value

Model plots paths (a list)